[#5] 메인 - 세팅 연결 진행
